Implement RV32C compressed instruction set
This extends the simulator to add support for the RV32C compressed
instructions extension. This extension does not add any instructions,
but provides shorter aliases for some 32-bit instructions. It also
makes 4-byte instructions starting on 2-byte boundaries valid.

Analogous to the existing code, the RV32C instruction parsing code is
generated from a table in `interp.in.rs`. This required a few small
changes in `build/cpu.rs`.

Extractors are added for the various fields specific to 16-bit
instruction encoding.

`rv32uc` is added to the ISA tests to test the decoding and emulation
code.

Gate the functionality behind a crate feature `rv32c`.